+++ GgafLib/include/jp/ggaf/lib/actor/DefaultCamera.h
@@ -11,8 +11,8 @@ namespace GgafLib {
 class DefaultCamera : public GgafDxCore::GgafDxCamera {
 
 public:
-    DefaultCamera(const char* prm_name, float prm_rad_fovX, float prm_dep);
-    DefaultCamera(const char* prm_name, float prm_rad_fovX);
+    DefaultCamera(const char* prm_name, double prm_rad_fovX, double prm_dep);
+    DefaultCamera(const char* prm_name, double prm_rad_fovX);
     DefaultCamera(const char* prm_name);
 
     virtual ~DefaultCamera(); //デストラクタ
